Extract unit test and refactor it to be ten times faster
Now runs in 0.065s instead of 0.70s
